pyproject.toml
[project]
name = "replearningbook"
version = "0.1.0"
description = "Files to run experiments for the book."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "openai>=1.0.0",
    "jax>=0.6.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "optax>=0.2.5",
    "pillow>=12.1.0",
    "pyright>=1.1.403",
    "requests>=2.32.4",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
